♻️ refactor contract namespace
