+++ src/Http/Controllers/ToolController.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace BBSLab\NovaFileManager\Http\Controllers\Pages;
+namespace BBSLab\NovaFileManager\Http\Controllers;
 
 use Illuminate\Routing\Controller;
 use Inertia\Inertia;
@@ -10,6 +10,11 @@ use Inertia\Response;
 
 class ToolController extends Controller
 {
+    /**
+     * The tool's Inertia component render
+     *
+     * @return \Inertia\Response
+     */
     public function __invoke(): Response
     {
         return Inertia::render('NovaFileManager', [
