+++ src/Http/Requests/DownloadFileRequest.php
@@ -7,10 +7,6 @@ namespace BBSLab\NovaFileManager\Http\Requests;
 use BBSLab\NovaFileManager\Rules\DiskExistsRule;
 use BBSLab\NovaFileManager\Rules\PathExistsInDiskRule;
 
-/**
- * @property-read ?string $disk
- * @property-read string $path
- */
 class DownloadFileRequest extends BaseRequest
 {
     public function rules(): array
