+++ src/ToolServiceProvider.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace BBSLab\NovaFileManager;
 
-use BBSLab\NovaFileManager\Contracts\FileManagerContract;
+use BBSLab\NovaFileManager\Contracts\Services\FileManagerContract;
 use BBSLab\NovaFileManager\Contracts\Filesystem\Upload\Uploader as UploaderContract;
 use BBSLab\NovaFileManager\Filesystem\Upload\Uploader;
 use BBSLab\NovaFileManager\Http\Middleware\Authorize;
